feat: change tasa_interes type from Decimal to float in ConfiguracionResponse and ConfiguracionActualizar schemas

+++ app/modules/mas/schemas/configuracion.py
@@ -1,13 +1,12 @@
 from pydantic import BaseModel
-from decimal import Decimal
 
 
 class ConfiguracionResponse(BaseModel):
     codigo_configuracion_pk: int
-    tasa_interes: Decimal
+    tasa_interes: float
 
     model_config = {"from_attributes": True}
 
 
 class ConfiguracionActualizar(BaseModel):
-    tasa_interes: Decimal
+    tasa_interes: float
